Refactor redux code
